# [Core] Optimize sampler get_logprobs